fix(challenge): urlsafe base64 encoding
This fixes issue with wrong encoding being used for the challenge.
According to the specs[^1] the challenge should be base64 url-safe
encoded. Until now, the package used std encoding which uses slightly
different set of characters. This was reported in the linked issue
and we also encountered the same issue when testing our implementation
of webauthn e2e using github.com/descope/virtualwebauthn.

The commit removes the custom type for challenge and instead passes
around the challenge url-safe base64 encoded right from the point of
creation. Because the string is fully valid it can be safely used
in json as well as stored in redis or other possible storage
implementations that might be used by the clients of the library.

I tried to keep the tests as close to the original as possible.
In some cases that required transforming the challenge back to bytes
considering the previously used padding to be part of the bytes
and then re-encoding into url-safe base64 string.

[^1]: https://www.w3.org/TR/webauthn-2/#dom-collectedclientdata-challenge

Fix: https://github.com/duo-labs/webauthn/issues/128