fix(webauthn): missing user display name from session (#116)
This includes the missing user display name value from the session data so if a relying party desires they can easily save the value during the finish registration stage for later to easily show the user. This is because the display name value should be set by the user and not the relying party.