feat(metadata)!: rework as a provider (#239)
This is a massive rework to how metadata is handled. This allows for very powerful interactions with the metadata including trust anchor validation, status report validation customization, control over if attestation types are validated against the metadata, exclusion of zero AAGUID's, ensuring a metadata entry for the authentication exists, etc.  In addition it allows the implementer to more easily control caching logic and implement locking and live refreshes of the metadata where necessary. It should reasonably be expected that anyone currently using the metadata package will have quite a bit of work to do. In the release we'll aim to include migration instructions but there is unfortunately it will require some additional work. 

Closes #77, Closes #154

BREAKING CHANGE: This change will require manual intervention from the implementer. Information is likely to be provided at a later date helping with the migrations required.